Try to fix project deletion edge case (#5534)
* Try to fix project deletion edge case

* Use feedback

* Use `create_demo_team` in Team deletion test

* Don't use raw `DELETE` queries
