+++ posthog/models/event.py
@@ -386,7 +386,6 @@ class Event(models.Model):
         return filtered_actions
 
     created_at: models.DateTimeField = models.DateTimeField(auto_now_add=True, null=True, blank=True)
-    objects: EventManager = EventManager.as_manager()  # type: ignore
     team: models.ForeignKey = models.ForeignKey(Team, on_delete=models.CASCADE)
     event: models.CharField = models.CharField(max_length=200, null=True, blank=True)
     distinct_id: models.CharField = models.CharField(max_length=200)
@@ -397,3 +396,5 @@ class Event(models.Model):
 
     # DEPRECATED: elements are stored against element groups now
     elements: models.JSONField = models.JSONField(default=list, null=True, blank=True)
+
+    objects: EventManager = EventManager.as_manager()  # type: ignore
